Squashed commit of the following:

commit de2029a2ede6562a8933e042be7f0b8d20033b05
Author: 詹伟伟 <weiwei.zhan@baishancloud.com>
Date:   Mon Jan 13 20:08:53 2020 +0800

    ORCHSYM-7712: add system and jvm cpu load to nodeSnapshots, and rounded to 3 decimals.

+++ nifi-nar-bundles/nifi-framework-bundle/nifi-framework/nifi-web/nifi-web-api/src/main/java/org/apache/nifi/web/api/dto/DtoFactory.java
@@ -224,6 +224,7 @@ import org.apache.nifi.web.util.AppTypeAssessor.AppType;
 import org.apache.nifi.util.VersionHelper;
 
 import javax.ws.rs.WebApplicationException;
+import java.math.BigDecimal;
 import java.text.Collator;
 import java.text.NumberFormat;
 import java.util.ArrayList;
@@ -3164,8 +3165,8 @@ public final class DtoFactory {
         // processors
         snapshot.setAvailableProcessors(sysDiagnostics.getAvailableProcessors());
         snapshot.setProcessorLoadAverage(sysDiagnostics.getProcessorLoadAverage());
-        snapshot.setSystemCpuLoad(sysDiagnostics.getSystemCpuLoad());
-        snapshot.setJvmProcessCpuLoad(sysDiagnostics.getJvmProcessCpuLoad());
+        snapshot.setSystemCpuLoad(BigDecimal.valueOf(sysDiagnostics.getSystemCpuLoad()));
+        snapshot.setJvmProcessCpuLoad(BigDecimal.valueOf(sysDiagnostics.getJvmProcessCpuLoad()));
 
         // threads
         snapshot.setDaemonThreads(sysDiagnostics.getDaemonThreads());
